Move test for basic Person to separate file
